Move Player and constants to players.py. Move Mob and constants to mobs.py
