+++ src/dungeon_controller.py
@@ -13,7 +13,9 @@ import sounds
 import tilemap
 import view
 from creatures import humanoids
-from creatures.humanoids import Player, Mob, collide_hit_rect_with_rect
+from creatures.humanoids import collide_hit_rect_with_rect
+from creatures.mobs import Mob
+from creatures.players import Player
 from items.item_manager import ItemManager
 from model import Obstacle, Groups, GameObject, Timer, \
     DynamicObject, Waypoint, Group, ConflictGroups
