+++ src/hud.py
@@ -5,7 +5,7 @@ import pygame as pg
 import images
 import mods
 import settings
-from creatures.humanoids import Player
+from creatures.players import Player
 from draw_utils import draw_text
 
 NO_SELECTION = -1
