+++ src/test/pygame_mock.py
@@ -3,6 +3,7 @@ from typing import List
 import pygame
 
 import abilities
+import creatures.mobs
 import images
 import model
 import sounds
@@ -55,5 +56,5 @@ def initialize_gameobjects(groups: model.Groups, timer: model.Timer) -> None:
     model.GameObject.initialize_gameobjects(groups)
     model.DynamicObject.initialize_dynamic_objects(timer)
     blank_screen = pygame.Surface((800, 600))
-    hmn.Mob.init_class(blank_screen)
+    creatures.mobs.Mob.init_class(blank_screen)
     abilities.initialize_classes(timer)
